Addressed code review feedback
Move OpenPgpMetadataExtractor into EncryptionDetails so metadata
construction and S2K annotation live in one place. Delete the util class
and merge its tests into EncryptionDetailsTest.

Make Decryptor safe for concurrent use by removing last-decrypt instance
state. Protected decrypt methods now return DecryptedData (stream plus
EncryptionDetails); unpack applies metadata from a new
decrypt(PGPEncryptedDataList) overload.

Raise default signing from SHA256 to SHA384 in Encryptor and
OpenPgpProfile.

Log warnings when Bouncy Castle reflection for
PGPPBEEncryptedData.keyData fails.

Add BouncyCastleReflectionCompatibilityTest as a build guard for that
internal field.
