+++ src/main/java/org/c02e/jpgpj/OpenPgpProfile.java
@@ -6,12 +6,12 @@ package org.c02e.jpgpj;
  */
 public enum OpenPgpProfile {
     /**
-     * RFC 4880 / gpg 2.2 era: AES128, MDC, SHA256 signing,
+     * RFC 4880 / gpg 2.2 era: AES128, MDC, SHA384 signing,
      * SHA512 iterated+salted S2K, ZLIB compression.
      */
     Classic,
     /**
-     * RFC 9580 / gpg 2.4+ era: AES256, AEAD-OCB v6, SHA256 signing,
+     * RFC 9580 / gpg 2.4+ era: AES256, AEAD-OCB v6, SHA384 signing,
      * Argon2 S2K when passphrase-encrypting, ZLIB compression.
      */
     Modern
